Small add

+++ synthetic_probability.py
@@ -75,8 +75,9 @@ class shift_and_log_scale_fn_helper(tf.Module):
         super(shift_and_log_scale_fn_helper, self).__init__(name=name)
         self.made = made
         self._made_variables = made.variables
+
     def __call__(self, x):
-        return tf.exp(-0.05*tf.norm(x, ord=2, axis=-1, keepdims=False)**2)[...,None,None] * self.made(x)
+        return tf.exp(-0.05*tf.norm(x, ord=2, axis=-1, keepdims=False)**2)[..., None, None] * self.made(x)
 
 
 class SimpleMAF(object):
@@ -134,7 +135,7 @@ class SimpleMAF(object):
             maf = tfb.MaskedAutoregressiveFlow(shift_and_log_scale_fn=shift_and_log_scale_fn)
             bijectors.append(maf)
 
-            if _permutations: # add the inverse permutation
+            if _permutations:  # add the inverse permutation
                 inv_perm = np.zeros_like(_permutations[i])
                 inv_perm[_permutations[i]] = np.arange(len(inv_perm))
                 bijectors.append(tfb.Permute(inv_perm.astype(np.int32)))
@@ -193,7 +194,7 @@ def prior_bijector_helper(prior_dict_list, name=None, **kwargs):
 
     """
     def uniform(a, b):
-        return tfb.Chain([tfb.Shift((a+b)/2), tfb.Scale((b-a)), tfb.Shift(-0.5), tfb.NormalCDF()])
+        return tfb.Chain([tfb.Shift(((a+b)/2.).astype(np.float32)), tfb.Scale((b-a)), tfb.Shift(-0.5), tfb.NormalCDF()])
 
     def normal(mu, sig):
         return tfb.Chain([tfb.Shift(mu), tfb.Scale(sig)])
